refactor(studio): extract orender service + mpv-overlay command modules
Phase 2, the largest command extraction. Moves two more groups out of main.rs,
no behaviour change (command names/signatures unchanged):

- commands/orender.rs — the orender launch + OS-service subsystem: 9 commands
  (status / install / uninstall / start / stop / restart, launch_orender,
  stop_orender, restart_pipewire_services) plus the ~17 private platform helpers
  they share (launch-spec resolution, run_command, systemd/SC glue) and the
  `OrenderLaunchSpec` / `OrenderServiceStatus` types. This whole subsystem was
  self-contained, so it moves as a unit; the helpers stay module-private.
- commands/mpv_overlay.rs — 8 commands: overlay trail prefs, visibility, and the
  energy heatmap (enabled / bands / colormap / custom stops).

Also drops the now-unused `std::{fs::File, process::*}` import from main.rs (only
the orender code used those).

main.rs drops from 1622 to 836 lines (2690 -> 836 across the Phase 2 PRs, ~69%).
Tauri app builds; fmt clean.
